requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
branca==0.5.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
cycler==0.11.0
discord.py==1.7.3
dotenv-python==0.0.1
folium==0.12.1.post1
fonttools==4.34.4
idna==3.3
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
networkx==2.8.5
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
polyline==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.11
yarl==1.7.2
